Revert "refactor: styles updated"
This reverts commit 95d97865e19dd103b8765264191591fbbcc10516.
