+++ packages/checkbox-group/package.json
@@ -22,14 +22,14 @@
     "test:cov": "yarn test --coverage"
   },
   "dependencies": {
-    "@jfront/ui-icons": "^0.1.1"
+    "@jfront/ui-icons": "^0.1.1",
+    "@jfront/ui-label": "^0.3.1"
   },
   "peerDependencies": {
     "react": "16.x",
     "react-dom": "16.x"
   },
   "devDependencies": {
-    "@jfront/ui-label": "^0.3.1",
     "@types/styled-components": "^5.0.1"
   }
 }
